Move survey_universe to empire module and call it earlier
