+++ default/python/AI/empire/colony_builders.py
@@ -2,18 +2,11 @@ from typing import Dict, List, Mapping, Sequence, Union
 
 import AIDependencies
 from common.fo_typing import PlanetId, SpeciesName
-from common.listeners import listener, register_post_handler, register_pre_handler
+from empire.survey_lock import survey_universe_lock
 from freeorion_tools import tech_is_complete
 from freeorion_tools.caching import cache_for_current_turn
-from freeorion_tools.lazy_initializer import InitializerLock
-
-_colony_builder_lock = InitializerLock("colony_builder")
-
-register_pre_handler("generateOrders", _colony_builder_lock.lock)
-register_post_handler("set_colony_builders", lambda *args: _colony_builder_lock.unlock())
 
 
-@listener
 def set_colony_builders(species_name: SpeciesName, yards: Sequence[PlanetId]):
     """
     Add planet where you can build colonies for species.
@@ -25,17 +18,17 @@ def set_colony_builders(species_name: SpeciesName, yards: Sequence[PlanetId]):
     empire_colonizers.setdefault(species_name, []).extend(yards)
 
 
-@_colony_builder_lock
+@survey_universe_lock
 def can_build_colony_for_species(species_name: Union[SpeciesName, str]):
     return species_name in get_colony_builders()
 
 
-@_colony_builder_lock
+@survey_universe_lock
 def get_colony_builder_locations(species_name: SpeciesName) -> List[PlanetId]:
     return get_colony_builders()[species_name]
 
 
-@_colony_builder_lock
+@survey_universe_lock
 def can_build_only_sly_colonies():
     """
     Return true if empire could build only SP_SLY colonies.
@@ -48,7 +41,7 @@ def can_build_only_sly_colonies():
     return list(get_colony_builders()) == ["SP_SLY"]
 
 
-@_colony_builder_lock
+@survey_universe_lock
 def get_colony_builders() -> Mapping[SpeciesName, List[PlanetId]]:
     """
     Return map from the species to list of the planet where you could build a colony ship with it.
