+++ default/python/AI/empire/growth_specials.py
@@ -1,6 +1,7 @@
 from typing import Dict, List, Mapping
 
 from common.fo_typing import PlanetId, SpecialName
+from empire.survey_lock import survey_universe_lock
 from freeorion_tools.caching import cache_for_current_turn
 
 
@@ -8,6 +9,7 @@ def set_growth_special(special: SpecialName, pid: PlanetId):
     _get_growth_specials().setdefault(special, []).append(pid)
 
 
+@survey_universe_lock
 def get_growth_specials() -> Mapping[SpecialName, List[PlanetId]]:
     """
     Return map from the species to list of the planet where you could build a ship with it.
