+++ default/python/AI/empire/pilot_rating.py
@@ -1,6 +1,7 @@
 from typing import Dict, Mapping, Sequence
 
 from common.fo_typing import PlanetId
+from empire.survey_lock import survey_universe_lock, survey_universe_lock2
 from freeorion_tools.caching import cache_for_current_turn
 
 
@@ -14,10 +15,12 @@ def set_pilot_rating_for_planet(pid: PlanetId, pilot_rating: float):
     _get_pilot_ratings()[pid] = pilot_rating
 
 
+@survey_universe_lock
 def get_rating_for_planet(pid: PlanetId) -> float:
     return get_pilot_ratings().get(pid, 0)
 
 
+@survey_universe_lock
 def get_pilot_ratings() -> Mapping[PlanetId, float]:
     return _get_pilot_ratings()
 
@@ -41,10 +44,12 @@ def _get_pilot_summary() -> _Summary:
     return _Summary()
 
 
+@survey_universe_lock2
 def best_pilot_rating() -> float:
     return _get_pilot_summary().best_pilot_rating
 
 
+@survey_universe_lock2
 def medium_pilot_rating() -> float:
     return _get_pilot_summary().medium_pilot_rating
